feat!: remove file uploads feature
