+++ package.json
@@ -62,9 +62,7 @@
   },
   "dependencies": {
     "@graphql-typed-document-node/core": "^3.2.0",
-    "cross-fetch": "^3.1.5",
-    "extract-files": "^9.0.0",
-    "form-data": "^3.0.0"
+    "cross-fetch": "^3.1.5"
   },
   "peerDependencies": {
     "graphql": "14 - 16"
@@ -75,7 +73,6 @@
     "@tsconfig/node16": "^1.0.3",
     "@types/body-parser": "^1.19.2",
     "@types/express": "^4.17.17",
-    "@types/extract-files": "^8.1.1",
     "@types/node": "^18.15.11",
     "@types/ws": "^8.5.4",
     "@typescript-eslint/eslint-plugin": "^5.57.1",
